+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Which is based on [Keep A Changelog](http://keepachangelog.com/)
 ## Changed
 
 - Prefix error message with ERROR 
+- Fix new line with clear input
 
 ## v1.0.0 - 2019-07-12